test: split flow and deployment tests in 2
